weatherapp/requirements.txt
aiogram==3.0.0b7
requests==2.31.0
python-dotenv==1.0.0